# docs/swagger.yaml
definitions:
  httpHandler.errorResponse:
    properties:
      error:
        type: string
    type: object
  schedule.CreateScheduleDTO:
    properties:
      duration:
        description: days
        type: integer
      name:
        type: string
      period:
        example: 1h30m
        type: string
      user_id:
        type: integer
    type: object
  schedule.CreateScheduleResponseDTO:
    properties:
      id:
        type: integer
    type: object
  schedule.NextTakingResponseDTO:
    properties:
      end_at:
        example: "2025-04-21T22:00:00Z"
        type: string
      id:
        type: integer
      name:
        type: string
      next_taking:
        example: "2025-04-21T08:00:00Z"
        type: string
      period:
        example: 1h30m
        type: string
    type: object
  schedule.ScheduleResponseDTO:
    properties:
      end_at:
        example: "2025-04-21T22:00:00Z"
        type: string
      id:
        type: integer
      name:
        type: string
      period:
        example: 1h30m
        type: string
      timetable:
        example:
        - "2025-04-21T08:00:00Z"
        items:
          type: string
        type: array
    type: object
info:
  contact: {}
paths:
  /next_taking:
    get:
      consumes:
      - application/json
      description: Возвращает данные о расписаниях на ближайший период
      parameters:
      - description: user id
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schedule.NextTakingResponseDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpHandler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpHandler.errorResponse'
      summary: Get next takings
      tags:
      - schedule
  /schedule:
    get:
      consumes:
      - application/json
      description: Возвращает данные о выбранном расписании с рассчитанным графиком
        приёмов на день
      parameters:
      - description: user id
        in: query
        name: user_id
        required: true
        type: integer
      - description: schedule id
        in: query
        name: schedule_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schedule.ScheduleResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpHandler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpHandler.errorResponse'
      summary: Get schedule
      tags:
      - schedule
    post:
      consumes:
      - application/json
      description: Создаёт новое расписание
      parameters:
      - description: schedule info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/schedule.CreateScheduleDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schedule.CreateScheduleResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpHandler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpHandler.errorResponse'
      summary: Create schedule
      tags:
      - schedule
  /schedules:
    get:
      consumes:
      - application/json
      description: Возвращает список идентификаторов существующих расписаний для указанного
        пользователя
      parameters:
      - description: user id
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: integer
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpHandler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpHandler.errorResponse'
      summary: Get user schedules
      tags:
      - schedule
swagger: "2.0"